+++ dev/mld/gen_all.py
@@ -1,0 +1,57 @@
+# Code used to generate the tiles for the MLD investigation
+
+import sys
+import os
+import logging
+
+# locals
+from mld_defs import MLD_DEFS
+
+from dbof.tiles import tile_utils
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s | %(levelname)s | %(message)s",
+    stream=sys.stdout,
+)
+
+def gen_tiles(timestamp:str='2012-11-09 12:00:00',
+        output_path:str=os.path.join(os.getenv('OS_OGCM'), 'LLC',
+        'Fronts', 'V4', '20121109_120000', 'tiles')) -> None:
+
+    # Generate folder
+    os.makedirs(output_path, exist_ok=True)
+
+    properties = ['density', 'temperature']
+    regions = ['tropical_pacific', 'gs_central', 
+                'gs_nc_1', 'gs_nc_2', 'gs_north_atlantic', 
+                'kuroshio', 'california_current']
+
+    # Loop me
+    for region in regions:
+        print('--------------------------------')
+        print(f"Processing region: {region}")
+        print('--------------------------------')
+        for prop in properties:
+            tile_utils.run(
+                i_rect=MLD_DEFS[region]['i_j'][0],
+                j_rect=MLD_DEFS[region]['i_j'][1],
+                timestamp=timestamp,
+                property=prop,
+                output=output_path)
+
+def main(flg):
+    flg = int(flg)
+
+    # Generate tiles
+    if flg == 1:
+        gen_tiles()
+
+# Command line
+if __name__ == '__main__':
+    if len(sys.argv) == 1:
+        flg = 1
+    else:
+        flg = sys.argv[1]
+
+    main(flg)
